lambda function refactor
