+++ frontend/src/pages/Login.jsx
@@ -3,7 +3,7 @@ import { useNavigate, Link } from 'react-router-dom';
 import { CognitoIdentityProviderClient, InitiateAuthCommand } from '@aws-sdk/client-cognito-identity-provider';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
-import { config } from '../congif';
+import { config } from '../config';
 
 
 const client = new CognitoIdentityProviderClient({ region: 'us-east-1' });
